+++ c8ylp/tcp_socket/tcp_server.py
@@ -46,7 +46,9 @@ class TCPHandler(socketserver.BaseRequestHandler):
 
         def handle_shutdown():
             # Force shutdown of any socket reads or writes
-            request.shutdown(socket.SHUT_RDWR)
+            # Note: fileno is set to -1 if the socket has been closed
+            if request.fileno() != -1:
+                request.shutdown(socket.SHUT_RDWR)
 
         self.server.web_socket_client.shutdown_request = handle_shutdown
 
@@ -81,6 +83,7 @@ class TCPHandler(socketserver.BaseRequestHandler):
                 if not data:
                     logging.debug("No data. Request will be closed")
                     self.server.web_socket_client.proxy_send_message = None
+                    self.server.web_socket_client.stop()
                     break
 
                 logging.debug("Writing data to ws: %s", data)